refactor: 사이드바 리스트 컴포넌트 분리
