+++ src/components/Sidebar/Sidebar.tsx
@@ -1,61 +1,20 @@
-import { useState } from 'react';
+import React from 'react';
 import styled from 'styled-components';
-import SidebarListItem from './SidebarListItem.tsx';
 import MapIcon from '../../assets/icons/map-pin-front-color.svg?react';
-import PlusIcon from '../../assets/icons/plus-front-clay.svg?react';
-import useLocations from '../../hooks/useLocation.tsx';
-import AddLocationModal from '../Modal/AddLocationModal.tsx';
+import SidebarList from './SidebarList';
+import { SIDEBAR_WIDTH } from '../../constans';
 
-interface SidebarProps {
-  selectedLocation: string | null;
-  onSelectLocation: (location: string) => void;
-}
-
-const Sidebar = ({ selectedLocation, onSelectLocation }: SidebarProps) => {
-  const { locations, addLocation, removeLocation } = useLocations();
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const handleAddLocation = (location: string) => {
-    addLocation(location);
-    setIsModalOpen(false);
-  };
-
+const Sidebar = () => {
   return (
-    <>
-      <Wrapper>
-        <TitleRow>
-          <StyledIcon>
-            <MapIcon />
-          </StyledIcon>
-          <TitleText>위치 목록</TitleText>
-        </TitleRow>
-
-        <AddRow onClick={() => setIsModalOpen(true)}>
-          <StyledIcon>
-            <PlusIcon />
-          </StyledIcon>
-          <TitleText>추가하기</TitleText>
-        </AddRow>
-
-        <LocationList>
-          {locations.map(location => (
-            <SidebarListItem
-              key={location}
-              location={location}
-              isSelected={selectedLocation === location}
-              onClick={() => onSelectLocation(location)}
-              onDelete={() => removeLocation(location)}
-            />
-          ))}
-        </LocationList>
-      </Wrapper>
-
-      <AddLocationModal
-        isOpen={isModalOpen}
-        onClose={() => setIsModalOpen(false)}
-        onConfirm={handleAddLocation}
-      />
-    </>
+    <Wrapper>
+      <TitleRow>
+        <StyledIcon>
+          <MapIcon />
+        </StyledIcon>
+        <TitleText>위치 목록</TitleText>
+      </TitleRow>
+      <SidebarList />
+    </Wrapper>
   );
 };
 
@@ -63,7 +22,6 @@ export default Sidebar;
 
 const Wrapper = styled.div`
   width: ${SIDEBAR_WIDTH};
-  width: 248px;
   height: 100vh;
   padding: 48px 16px;
   display: flex;
@@ -87,22 +45,10 @@ const TitleText = styled.div`
   font-weight: 700;
 `;
 
-const AddRow = styled(TitleRow)`
-  cursor: pointer;
-`;
-
 const StyledIcon = styled.div`
   width: 40px;
   height: 40px;
   display: flex;
   align-items: center;
   justify-content: center;
-`;
-
-const LocationList = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 8px;
-  width: 100%;
 `;
