+++ src/constans/index.ts
@@ -1,0 +1,1 @@
+export const SIDEBAR_WIDTH = '248px';
